pyproject.toml
[project]
name = "mindsetbackend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
authors = [
{ name = "Amirehsan Ahmadzadeh" }
]
license = {text = "MIT"}
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-migrate>=4.0.7",
    "typing-extensions>=4.12.2",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "dotenv>=0.9.9",
    "flasgger>=0.9.7.1",
    "alembic>=1.17.0",
    "rich-click>=1.9.3",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "pytest-cov>=7.0.0",
]


[project.optional-dependencies]
dev = [
"pytest>=8.3.0",
"pytest-cov>=5.0.0",
"ruff>=0.6.0",
"httpx>=0.27.0",
"pyright>=1.1.386",
]


[tool.uv]
# uv-specific settings can go here if needed later


[tool.ruff]
line-length = 100
target-version = "py313"


[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "PIE", "Q", "T20"]
ignore = [
# Example: "D" for pydocstyle if later enabled
]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"


[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"                  # "basic" if this is too noisy at first
include = ["app"]                            # your source root(s)
exclude = [".venv", ".git", "build", "dist", ".pytest_cache", ".ruff_cache"]
venvPath = "."                               # let pyright find .venv next to project
venv = ".venv"


[tool.pytest.ini_options]
addopts = "-q"
